requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
pydantic>=1.10
python-dotenv>=1.0
openai>=1.0
pytest>=7.0
authlib>=1.2.0
httpx>=0.24.0
itsdangerous>=2.1.0
starlette>=0.27.0
